That's it, tag the ReactOS 0.3.7 release

+++ reactos/lib/rtl/res.c
@@ -60,7 +60,7 @@ static int is_data_file_module( PVOID BaseAddress )
  */
 int push_language( USHORT *list, ULONG pos, WORD lang )
 {
-    ULONG i;
+    int i;
     for (i = 0; i < pos; i++) if (list[i] == lang) return pos;
     list[pos++] = lang;
     return pos;
